Move compiler utils that needs to CLI to the Commons package
