+++ commons/build.gradle
@@ -27,6 +27,9 @@ description = 'Ballerina - GraphQL Native'
 dependencies {
     checkstyle project(':checkstyle')
     checkstyle "com.puppycrawl.tools:checkstyle:${checkstylePluginVersion}"
+
+    implementation group: 'org.ballerinalang', name: 'ballerina-lang', version: "${ballerinaLangVersion}"
+    implementation group: 'org.ballerinalang', name: 'ballerina-parser', version: "${ballerinaLangVersion}"
 }
 
 def excludePattern = '**/module-info.java'
